models: Add default values to the lat/lng fields
This allows us to hide the input for these fields from the UI and fill
them programatically on the front end. If this fails for whatever reason
we don't want the user to have to be blocked by looking up the lat/lng
values so set them to 99 which is an invalid coordinate for future
geocoding.