widgets: add-object-modal: Improve the state change handling
Improve the state change handling specifically resuming the old state of
the previous model. Make sure that if a fk has been added and then
selected that it is set as the new fk value for the original object.
Also make sure that this is visually indicated.

Also Add fall-back for tinyMCE failure (note should have been a separate
commit)