widgets: add-object-modal: close handler and fallback field type
Make sure when BS closes the modal that we hook into the close handler
on the Vue side of things.

If we haven't got a specific handler for the type of field then just
fall back to to a text input.

Also Add a hint about required fields. This is mainly to reinforce the
fact that those not marked with the required really are optional.